fix(number_formatter): avoid floating point precision errors
This commit formats the string of numbers by adding the thousands
separator to every third character from the end of each whole number.

The fundamental way by which Python handles floating point arithmetics
(https://docs.python.org/3/library/decimal.html) causes precision errors
on larger numbers (e.g., `9999999999.9999999999`). Hence, the changes in
this commit avoid using `float()` and `int()` functions for creating and
formatting the actual output strings. Note that both of those functions
are still used for verification and control proposes: when evaluating
whether the string is a number and when comparing it to other numbers.

#5